refactor(slideshow): simplify StepHardSkills and StepSoftSkills components by utilizing custom hooks
- Removed local state management for selected and expanded skills in StepHardSkills and StepSoftSkills components.
- Integrated useHardSkills and useSoftSkills hooks to handle skill selection and management.
- Updated types to include softSkills and hardSkills in the Slide interface for better state representation.
